+++ src/main/java/me/drex/instantfeedback/worldgen/ModVegetationFeatures.java
@@ -10,21 +10,18 @@ import net.minecraft.core.HolderGetter;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.data.worldgen.BootstrapContext;
 import net.minecraft.data.worldgen.features.FeatureUtils;
-import net.minecraft.data.worldgen.placement.PlacementUtils;
 import net.minecraft.data.worldgen.placement.TreePlacements;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.Identifier;
 import net.minecraft.util.random.WeightedList;
 import net.minecraft.util.valueproviders.ConstantInt;
 import net.minecraft.world.level.block.Blocks;
-import net.minecraft.world.level.block.LeavesBlock;
 import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.level.levelgen.blockpredicates.BlockPredicate;
 import net.minecraft.world.level.levelgen.feature.ConfiguredFeature;
 import net.minecraft.world.level.levelgen.feature.Feature;
 import net.minecraft.world.level.levelgen.feature.WeightedPlacedFeature;
+import net.minecraft.world.level.levelgen.feature.configurations.BlockPileConfiguration;
 import net.minecraft.world.level.levelgen.feature.configurations.RandomFeatureConfiguration;
-import net.minecraft.world.level.levelgen.feature.configurations.RandomPatchConfiguration;
 import net.minecraft.world.level.levelgen.feature.configurations.SimpleBlockConfiguration;
 import net.minecraft.world.level.levelgen.feature.configurations.TreeConfiguration;
 import net.minecraft.world.level.levelgen.feature.featuresize.ThreeLayersFeatureSize;
@@ -41,7 +38,6 @@ import java.util.OptionalInt;
 public class ModVegetationFeatures {
 
     public static final ResourceKey<ConfiguredFeature<?, ?>> PATCH_PALE_PUMPKIN = ResourceKey.create(Registries.CONFIGURED_FEATURE, Identifier.fromNamespaceAndPath(InstantFeedback.MOD_ID, "patch_pale_pumpkin"));
-    public static final ResourceKey<ConfiguredFeature<?, ?>> PILE_PALE_LEAVES = ResourceKey.create(Registries.CONFIGURED_FEATURE, Identifier.fromNamespaceAndPath(InstantFeedback.MOD_ID, "pile_pale_leaves"));
     public static final ResourceKey<ConfiguredFeature<?, ?>> PALE_VEGETATION = ResourceKey.create(Registries.CONFIGURED_FEATURE, Identifier.fromNamespaceAndPath(InstantFeedback.MOD_ID, "pale_vegetation"));
     public static final ResourceKey<ConfiguredFeature<?, ?>> FALLEN_PALE_OAK_CREAKING = ResourceKey.create(Registries.CONFIGURED_FEATURE, Identifier.fromNamespaceAndPath(InstantFeedback.MOD_ID, "fallen_pale_oak_creaking"));
     public static final ResourceKey<ConfiguredFeature<?, ?>> PALE_GARDEN_VEGETATION = ResourceKey.create(Registries.CONFIGURED_FEATURE, Identifier.fromNamespaceAndPath(InstantFeedback.MOD_ID, "pale_garden_vegetation"));
@@ -54,44 +50,32 @@ public class ModVegetationFeatures {
 
         FeatureUtils.register(
             bootstrapContext,
-            ModVegetationFeatures.PATCH_PALE_PUMPKIN,
-            Feature.RANDOM_PATCH,
-            FeatureUtils.simplePatchConfiguration(
-                Feature.SIMPLE_BLOCK, new SimpleBlockConfiguration(new WeightedStateProvider(
+            PATCH_PALE_PUMPKIN,
+            Feature.BLOCK_PILE,
+            new BlockPileConfiguration(
+                new WeightedStateProvider(
                     WeightedList.<BlockState>builder()
                         .add(ModBlocks.PALE_PUMPKIN.defaultBlockState(), 20)
                         .add(ModBlocks.CARVED_PALE_PUMPKIN.defaultBlockState().setValue(CarvedPalePumpkinBlock.FACING, Direction.NORTH), 1)
                         .add(ModBlocks.CARVED_PALE_PUMPKIN.defaultBlockState().setValue(CarvedPalePumpkinBlock.FACING, Direction.EAST), 1)
                         .add(ModBlocks.CARVED_PALE_PUMPKIN.defaultBlockState().setValue(CarvedPalePumpkinBlock.FACING, Direction.SOUTH), 1)
                         .add(ModBlocks.CARVED_PALE_PUMPKIN.defaultBlockState().setValue(CarvedPalePumpkinBlock.FACING, Direction.WEST), 1)
-                )), List.of(Blocks.GRASS_BLOCK),
-                24
-            )
-        );
-
-        FeatureUtils.register(
-            bootstrapContext, PILE_PALE_LEAVES, Feature.RANDOM_PATCH, FeatureUtils.simplePatchConfiguration(
-                Feature.SIMPLE_BLOCK,
-                new SimpleBlockConfiguration(BlockStateProvider.simple(Blocks.PALE_OAK_LEAVES.defaultBlockState().setValue(LeavesBlock.PERSISTENT, true))),
-                List.of(Blocks.GRASS_BLOCK, Blocks.PALE_OAK_LEAVES)
+                )
             )
         );
 
         FeatureUtils.register(
             bootstrapContext,
             PALE_VEGETATION,
-            Feature.RANDOM_PATCH,
-            new RandomPatchConfiguration(164, 16, 8, PlacementUtils.filtered(
-                Feature.SIMPLE_BLOCK,
-                new SimpleBlockConfiguration(new WeightedStateProvider(
+            Feature.SIMPLE_BLOCK,
+            new SimpleBlockConfiguration(
+                new WeightedStateProvider(
                     WeightedList.<BlockState>builder()
                         .add(ModBlocks.PALE_ROSE.defaultBlockState(), 1)
                         .add(ModBlocks.PALE_BUSH.defaultBlockState(), 5)
                         .add(ModBlocks.TALL_PALE_BUSH.defaultBlockState(), 5)
                         .add(Blocks.PALE_MOSS_CARPET.defaultBlockState(), 10)
-                )),
-                BlockPredicate.allOf(BlockPredicate.ONLY_IN_AIR_PREDICATE, BlockPredicate.matchesBlocks(Direction.DOWN.getUnitVec3i(), List.of(Blocks.GRASS_BLOCK)))
-            )
+                )
             )
         );
 
@@ -115,7 +99,14 @@ public class ModVegetationFeatures {
             bootstrapContext,
             PALE_GARDEN_VEGETATION,
             Feature.RANDOM_SELECTOR,
-            new RandomFeatureConfiguration(List.of(new WeightedPlacedFeature(paleOakCreakingChecked, 0.2F), new WeightedPlacedFeature(paleOakChecked, 0.8F), new WeightedPlacedFeature(fallenPaleOakCreakingChecked, 0.8F)), paleOakChecked)
+            new RandomFeatureConfiguration(
+                List.of(
+                    new WeightedPlacedFeature(paleOakCreakingChecked, 0.2F),
+                    new WeightedPlacedFeature(paleOakChecked, 0.8F),
+                    new WeightedPlacedFeature(fallenPaleOakCreakingChecked, 0.8F)
+                ),
+                paleOakChecked
+            )
         );
 
     }